# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Booking Clone API",
    "description": "REST API za aplikaciju za rezervaciju apartmana.\n\nTri uloge: **USER** rezerviše, **HOST** izdaje i vodi svoje apartmane,\n**ADMIN** upravlja šifarnicima.\n\n**Autentifikacija.** Prijava vraća *access token* koji traje 15 minuta i putuje\nu zaglavlju `Authorization: Bearer <token>`. Uz njega ide i *refresh token*, ali\non stoji u `httponly` kolačiću i ne čita se iz JavaScript-a; `POST /auth/refresh`\nga troši i izdaje nov par, pa jedan refresh token vredi tačno jednom.\n\nU Swagger UI-ju se prijavljujete dugmetom **Authorize**, unosom email adrese i\nlozinke.\n\n**Greške.** Svaka greška ima isti oblik: `detail` je rečenica na engleskom,\n`code` je stabilna oznaka koju frontend prevodi na jezik korisnika, a `params`\nnosi vrednosti koje ulaze u prevedenu rečenicu. Zahvaljujući tome poruka o\ngrešci nikad nije sklopljena od engleskih parčića unutar srpske rečenice.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/stats/host/earnings": {
      "get": {
        "tags": [
          "stats"
        ],
        "summary": "Zarada domaćina po mesecima ili godinama",
        "description": "Earnings of the signed in host, one row per month or per year.\n\nThe adding up happens in Python on purpose. Grouping by month in SQL means\n`strftime` on SQLite and `date_trunc` on Postgres, so the query would have\nto be rewritten the day the database changes. For the number of bookings a\nsingle host has, reading them and summing them here costs nothing.",
        "operationId": "get_host_earnings_stats_host_earnings_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "group_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "month",
              "title": "Group By"
            }
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "name": "apartment_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Apartment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EarningsPointDto"
                  },
                  "title": "Response Get Host Earnings Stats Host Earnings Get"
                }
              }
            }
          },
          "400": {
            "description": "Zahtev je neispravan, npr. datum odjave pre datuma prijave.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Niste prijavljeni ili je token istekao.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Prijavljeni ste, ali nemate pravo na ovu radnju.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/stats/host/by-apartment": {
      "get": {
        "tags": [
          "stats"
        ],
        "summary": "Udeo svakog apartmana u zaradi",
        "description": "How the same money splits over the host's apartments, with percentages.\n\nAn apartment that earned nothing in the period is left out: a slice of zero\nis not a slice, it is only a name taking up room in the legend.",
        "operationId": "get_host_earnings_by_apartment_stats_host_by_apartment_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "group_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "month",
              "title": "Group By"
            }
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "name": "apartment_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Apartment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApartmentShareDto"
                  },
                  "title": "Response Get Host Earnings By Apartment Stats Host By Apartment Get"
                }
              }
            }
          },
          "400": {
            "description": "Zahtev je neispravan, npr. datum odjave pre datuma prijave.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Niste prijavljeni ili je token istekao.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Prijavljeni ste, ali nemate pravo na ovu radnju.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApartmentShareDto": {
        "properties": {
          "apartment_id": {
            "type": "integer",
            "title": "Apartment Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "total": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total"
          },
          "nights": {
            "type": "integer",
            "title": "Nights"
          },
          "reservations": {
            "type": "integer",
            "title": "Reservations"
          },
          "share_percent": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Share Percent"
          },
          "is_deleted": {
            "type": "boolean",
            "title": "Is Deleted"
          }
        },
        "type": "object",
        "required": [
          "apartment_id",
          "title",
          "total",
          "nights",
          "reservations",
          "share_percent",
          "is_deleted"
        ],
        "title": "ApartmentShareDto"
      },
...
      "EarningsPointDto": {
        "properties": {
          "period": {
            "type": "string",
            "title": "Period"
          },
          "total": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total"
          },
          "nights": {
            "type": "integer",
            "title": "Nights"
          },
          "reservations": {
            "type": "integer",
            "title": "Reservations"
          }
        },
        "type": "object",
        "required": [
          "period",
          "total",
          "nights",
          "reservations"
        ],
        "title": "EarningsPointDto"
      },
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Rečenica na engleskom, čitljiva i bez prevoda."
          },
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Stabilna oznaka greške koju frontend prevodi."
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params",
            "description": "Vrednosti koje ulaze u prevedenu poruku, npr. `max_guests`."
          }
        },
        "type": "object",
        "required": [
          "detail",
          "code"
        ],
        "title": "ErrorResponse",
        "description": "The one shape every failed request comes back in.\n\nBuilt by `ExceptionHandlers`, which is the only place that knows what an\nerror looks like on the wire. Declaring it here as well is what puts it in\nthe specification, where it is otherwise invisible.",
        "example": {
          "code": "too_many_guests",
          "detail": "This apartment allows at most 4 guests",
          "params": {
            "max_guests": 4
          }
        }
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/login"
          }
        }
      }
    }
  },
...
}